Migrate from styled components to scss modules
