+++ components/Subscribe/Subscribe.module.scss
@@ -1,0 +1,33 @@
+.form {
+  display: flex;
+  gap: 0.75rem;
+}
+
+.label {
+  position: relative;
+}
+
+.input {
+  font-size: 0.875rem;
+  line-height: 1.25rem;
+  padding: 0.75rem;
+  border-radius: 0.375rem;
+  border: 1px solid #d2d4dc;
+  min-width: 400px;
+  padding-right: 100px;
+}
+
+.button {
+  line-height: 1.25rem;
+  padding: 0.5rem 0.75rem;
+  border-radius: 0.375rem;
+  color: white;
+  background: black;
+  border: 1px solid black;
+  text-transform: uppercase;
+  font-size: 0.75rem;
+  font-weight: 600;
+  position: absolute;
+  top: 0.25rem;
+  right: 0.25rem;
+}
