+++ components/index.ts
@@ -1,1 +1,1 @@
-export { default as Subscribe } from "./Subscribe";
+export { default as Subscribe } from "./Subscribe/Subscribe";
