+++ package.json
@@ -11,16 +11,13 @@
   "dependencies": {
     "next": "12.1.5",
     "react": "18.1.0",
-    "react-dom": "18.1.0",
-    "styled-components": "^5.3.5"
+    "react-dom": "18.1.0"
   },
   "devDependencies": {
     "@axe-core/react": "^4.4.4",
     "@types/node": "17.0.29",
     "@types/react": "18.0.8",
     "@types/react-dom": "18.0.0",
-    "@types/styled-components": "^5.1.26",
-    "babel-plugin-styled-components": "^2.0.7",
     "eslint": "8.14.0",
     "eslint-config-next": "12.1.5",
     "sass": "^1.53.0",
