+++ styles/base/_global.scss
@@ -16,7 +16,7 @@ html,
 body {
   margin: 0;
   padding: 0;
-  font-family: "Work Sans", -apple-system, BlinkMacSystemFont, avenir next,
+  font-family: "Neufile Grotesk", -apple-system, BlinkMacSystemFont, avenir next,
     avenir, segoe ui, helvetica neue, helvetica, Cantarell, Ubuntu, roboto, noto,
     arial, sans-serif;
   -webkit-font-smoothing: antialiased;
