--- templates/Home/index.ts
@@ -1,2 +1,0 @@
-export * from './Home';
-export { default } from './Home';
