--- templates/index.ts
@@ -1,3 +1,0 @@
-import Home from "./Home/Home";
-
-export { Home };
